+++ src/core/http/queries.js
@@ -1,6 +1,6 @@
 import API from './api-settings';
 
-const QueriesConfiguration = (config) => {
+const QueriesConfiguration = config => {
   const { base } = config || {};
   const url = base || API.base;
